requirements.txt
atlassian-python-api>=3.20.1
